Modularizacion del proyecto y optimizacion del codigo Albert
