+++ Sistemas de ecuaciones/ecuacion.py
@@ -1,7 +1,4 @@
 import random
-import numpy as np
-import matplotlib.pyplot as plt
-from mpl_toolkits.mplot3d import axes3d
 
 class Ecuacion:
     def __init__(self):
@@ -10,13 +7,8 @@ class Ecuacion:
         self.filas = 0
         self.columnas = 0
 
-    def __str__(self):
-        return self.vector
-    
-    def __str__(self):
-        return self.matriz
 
-    def generar_matriz(self, filas = 3, columnas = 3, value = None):
+    def generar_sistema(self, filas = 3, columnas = 3, value = None):
         self.filas = filas
         self.columnas = columnas
         numero_b = int(input('Ingresa el rango bajo del numero =>'))
@@ -35,45 +27,3 @@ class Ecuacion:
             self.vector[i] = value
 
         return self.matriz, self.vector
-    
-ecuacion = Ecuacion()
-
-matriz, vector = ecuacion.generar_matriz()
-print(vector)
-print(matriz)
-
-class Graficacion:
-    def __init__(self):
-        self.punto=np.linalg.solve(matriz,vector)
-
-    def __str__(self):
-
-        return self.punto
-
-fig = plt.figure()
-ax = fig.add_subplot(111, projection='3d')
-
-x, y = np.linspace(-10, 10, 100), np.linspace(-10, 10, 100)
-X, Y = np.meshgrid(x, y)
-Z1 = (vector[0] - matriz[0][0]*X - matriz[0][1]*Y) / matriz[0][2]
-Z2 = (vector[1] - matriz[1][0]*X - matriz[1][1]*Y) / matriz[1][2]
-Z3 = (vector[2] - matriz[2][0]*X - matriz[2][1]*Y) / matriz[2][2]
-ax.plot_surface(X, Y, Z1, alpha=0.5)
-ax.plot_surface(X, Y, Z2, alpha=0.5)
-ax.plot_surface(X, Y, Z3, alpha=0.5)
-
-ax.scatter(Graficacion().punto[0], Graficacion().punto[1], Graficacion().punto[2], color='red')
-
-
-ax.set_xlabel('X')
-ax.set_ylabel('Y')
-ax.set_zlabel('Z')
-
-plt.show()
-
-   
-
-
-    
-
-
